rename package and artifact ids from documentdb to cosmosdb
