+++ .release-automation/config.json
@@ -6,7 +6,7 @@
     "azure-spring-boot",
     "azure-spring-boot-starter",
     "azure-active-directory-spring-boot-starter",
-    "azure-documentdb-spring-boot-starter",
+    "azure-cosmosdb-spring-boot-starter",
     "azure-storage-spring-boot-starter",
     "azure-keyvault-secrets-spring-boot-starter",
     "azure-mediaservices-spring-boot-starter",
