+++ azure-spring-boot-samples/azure-cosmosdb-spring-boot-sample/build.gradle
