+++ azure-spring-boot-samples/azure-cosmosdb-spring-boot-sample/pom.xml
@@ -10,10 +10,10 @@
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>azure-documentdb-spring-boot-sample</artifactId>
+    <artifactId>azure-cosmosdb-spring-boot-sample</artifactId>
     <packaging>jar</packaging>
 
-    <name>Azure Document DB Spring Boot Starter Sample</name>
+    <name>Azure Cosmos DB Spring Boot Starter Sample</name>
     <description>Sample project for Azure Document DB Starter</description>
     <url>https://github.com/Microsoft/azure-spring-boot</url>
