+++ azure-spring-boot-samples/azure-cosmosdb-spring-boot-sample/src/main/java/sample/documentdb/DocumentDbSampleApplication.java
