+++ azure-spring-boot-samples/azure-cosmosdb-spring-boot-sample/src/main/resources/application.properties
