+++ azure-spring-boot-starters/azure-cosmosdb-spring-boot-starter/pom.xml
@@ -11,9 +11,9 @@
         <relativePath>../../azure-spring-boot-parent/pom.xml</relativePath>
     </parent>
 
-    <artifactId>azure-documentdb-spring-boot-starter</artifactId>
+    <artifactId>azure-cosmosdb-spring-boot-starter</artifactId>
 
-    <name>Azure Document DB Spring Boot Starter</name>
+    <name>Azure Cosmos DB Spring Boot Starter</name>
     <description>Spring Boot Starter for Azure Document DB service</description>
     <url>https://github.com/Microsoft/azure-spring-boot</url>
