+++ azure-spring-boot-starters/pom.xml
@@ -21,7 +21,7 @@
     <modules>
         <module>azure-spring-boot-starter</module>
         <module>azure-active-directory-spring-boot-starter</module>
-        <module>azure-documentdb-spring-boot-starter</module>
+        <module>azure-cosmosdb-spring-boot-starter</module>
         <module>azure-keyvault-secrets-spring-boot-starter</module>
         <module>azure-mediaservices-spring-boot-starter</module>
         <module>azure-servicebus-spring-boot-starter</module>
