+++ azure-spring-boot/src/main/resources/META-INF/spring.factories
@@ -2,8 +2,8 @@ org.springframework.boot.env.EnvironmentPostProcessor=com.microsoft.azure.spring
 com.microsoft.azure.keyvault.spring.KeyVaultEnvironmentPostProcessor
 org.springframework.context.ApplicationContextInitializer=\
 com.microsoft.azure.spring.autoconfigure.aad.YamlFileApplicationContextInitializer
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.microsoft.azure.spring.autoconfigure.documentdb.DocumentDBAutoConfiguration,\
-com.microsoft.azure.spring.autoconfigure.documentdb.DocumentDbRepositoriesAutoConfiguration,\
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.microsoft.azure.spring.autoconfigure.cosmosdb.DocumentDBAutoConfiguration,\
+com.microsoft.azure.spring.autoconfigure.cosmosdb.DocumentDbRepositoriesAutoConfiguration,\
 com.microsoft.azure.spring.autoconfigure.gremlin.GremlinAutoConfiguration,\
 com.microsoft.azure.spring.autoconfigure.gremlin.GremlinRepositoriesAutoConfiguration,\
 com.microsoft.azure.spring.autoconfigure.mediaservices.MediaServicesAutoConfiguration,\
