+++ azure-spring-boot/src/test/java/com/microsoft/azure/spring/autoconfigure/cosmosdb/DocumentDBAutoConfigurationTest.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See LICENSE in the project root for
  * license information.
  */
-package com.microsoft.azure.spring.autoconfigure.documentdb;
+package com.microsoft.azure.spring.autoconfigure.cosmosdb;
 
 
 import com.microsoft.azure.documentdb.ConnectionPolicy;
